requirements.txt
boto3==1.24.24
botocore==1.27.24
Flask==2.1.2
matplotlib==3.5.2
moto==3.1.16
numpy
pandas
python-dateutil==2.8.1
PyYAML==6.0
redshift-connector
reportlab
sqlparse==0.4.2
tabulate==0.8.10
tqdm==4.59.0
maya
coverage
pre-commit
black
flake8
pytest
